requirements.txt
amqp==2.4.0
appnope==0.1.0
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.8.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
cssselect==1.1.0
decorator==4.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.18
django-crispy-forms==1.6.1
django-filter==2.2.0
django-formtools==2.1
django-import-export==1.0.0
django-redis==4.9.0
django-reversion==2.0.13
djangorestframework==3.10.3
et-xmlfile==1.0.1
future==0.15.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
httplib2==0.9.2
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.13.2
kombu==4.2.2.post1
lxml==4.4.1
mysqlclient==1.4.4
odfpy==1.4.0
openpyxl==2.6.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
PyMySQL==0.9.3
pyquery==1.4.0
pytz==2018.9
PyYAML==5.1.2
qiniu==7.2.2
redis==2.10.6
requests==2.21.0
serializers==0.2.4
simplegeneric==0.8.1
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
tablib==0.13.0
traitlets==4.3.2
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
xlrd==1.2.0
xlwt==1.3.0
